requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv
pytz==2025.2
redis==6.2.0
requests==2.32.4
six==1.17.0
spotipy==2.25.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn
